feat: MarketStatic 평균금액 필드 double -> long 변경

+++ src/main/java/eureca/capstone/project/batch/market_statistic/domain/MarketStatistic.java
@@ -23,7 +23,7 @@ public class MarketStatistic extends BaseEntity {
     private Long statisticsId;
 
     @Column(name = "average_price")
-    private double averagePrice;
+    private long averagePrice;
 
     @Column(name = "transaction_amount")
     private long transactionAmount;
